use traits
